LU-6566 mdt: optimize sub_stripe lock/unlock
In mdt_reint_setattr, it should only lock sub_stripes
for certain attributes, for other attributes, it only
needs to set on the master object.

In lod_object_unlock, since it already know the lock
handle, it can release the lock directly, instead of
load and holding the sub-stripe object.

Signed-off-by: wang di <di.wang@intel.com>
Change-Id: Id9267f3619b036ce943a1df6ae163ea49a0fbc25