LU-6562 tests: sanity-scrub parse 'prior_updated'
In sanity-scrub tests 4a, 4b, and 4c, the test is parsing the
osd-ldiskfs.*MDT*.io_scrub values for 'sf_items_updated_prior'.
The parameter these test acutally want to parse for in oi_scrub
is 'prior_updated'.

There are some additional modifications to clean up the tests
and to catch errors including adding checks for and calls to error
for some functions that return an error.

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I94be02e2a8e6a9fc3902a57f1ff7c9010e493504