LU-5092 nodemap: transfer nodemaps between MGS, MDTs, and OSTs
This creates a new config lock on the MGS for transfering the
nodemaps. The target MGCs can enqueue this lock, which is then
revoked by the MGS when the nodemaps are modifed. The MGCs then issue
a read config RPC on the nodemap config lock, which causes the MGS to
transfer the nodemap index file to the MGCs. The MGCs then process
the index file, similarly to the way the MGS processes it on start.

Test-Parameters: fortestonly
Signed-off-by: Kit Westneat <kit.westneat@gmail.com>
Change-Id: I99a34773298484f70a912761f5831b75196d41d8