LU-6475 mdt: race between open and migrate
During intent open, if it founds the parent has been
migrated to another MDT, it should retry the open
request with the new object, so it needs to keep the
old object in the orphan list, which will be cleanup
during next recovery. Note: if the client still using
the old FID after next recovery, it will return -ENOENT
for the application.

Reset parent FID on OST object after migration, which
also make sanity-lfsck.sh 15c obsolete, because that
test suppose to fix the parent FID of the OST object
after migration.

And also enqueue the lease lock of the migrating file,
then compare the lease before migration to make sure
no other clients open the file at the same time.

Add sanityn.sh 80b to test race between open/stat and
migration.

Signed-off-by: wang di <di.wang@intel.com>
Change-Id: Iad3e5006079e9f72b338680e17448c39c6d1af98