LU-5407 tests: Error message for replay-single 58b and 58c
In the replay-single test suite, tests 58b and 58c set an attribute,
retrieve the attribute and check that the original message matches
what was set. If large extended attributes is enabled, the attribute
written can be very long. In case of error, we only print the first
few charaters of the message that was set.

Also, call error if there is a problem when the client mounts the
file system on $MOUNT2.

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I8c51ab04d906d74d6eac5a65420203936184ffe4