LU-6625 llite: symlink() sets wrong errno, ENOENT
According to the online-manual symlink() should set errno EEXIST when
specified link-path already exists. but it actually sets ENOENT when
the last ward of the link path is "/".

This patch modifies a if-statement in ll_lookup_nd for client to look
up MDS properly and for symlink() to set errno EEXIST

Signed-off-by: Hiroya Nozaki <nozaki.hiroya@jp.fujitsu.com>
Change-Id: Iec603139d2ebebea339f185a481bd8d2a4074deb