LU-6556 obdclass: re-allow catalog loopback
Since patch for LU-4528 a LLOG catalog is no longer allowed to
loopback. This is a regression and it can also cause catalog
corruption (grow behind max-size/records) upon upgrading if
catalog has already looped-back.

This patch reintroduces catalog loopback capability, and also
introduces a new test to check it.

Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>
Change-Id: Ife9a452199895ed9d9f43eb9fdeeac15322e272a