LU-6271 ldlm: cleanup locks before activating import
If the thread doing IO doesn't reach cl_io_unlock in time,
an extent lock still in use will be skipped during
eviction cleanup on the client.
That lock could be cancelled and freed later when its not
in use anymore.
However if the import becomes active before the cancellation
of the stale lock, another thread could enqueue and get a
new lock covering the same extent, and starts to queue
dirty pages for the object. This could lead to a LASSERT
failure later when we cancel the old stale lock.

We should wait for the extent locks in the cleanup stage
and then reactivate import.

Grouplocks are special. We just discard the pages and skip
them during cleanup so we can proceed and reject
any subsequent IO from the grouplock fd, as the grouplock
is stale after the eviction anyway.

Leftover grouplocks are still released manually, by calling
ioctl or closing the fd.

Signed-off-by: Li Dongyang <dongyang.li@anu.edu.au>
Change-Id: Ia622e7d150521db07673f83fac9fdb9674ef3b85