LU-6069 osd: mark directory's inode dirty
at insert, to transfer new i_size to buffer cache.
ext4 does this late when mtime is being refreshed,
so ext4_append() does not on its own.
basically this change double the number of calls
to ldiskfs_mark_inode_dirty() for file/directory
creations and this implies performance regression.
probably we should be putting the inode on a list
and mark it dirty once a transaction or even once
a transaction group..

Test-Parameters: fortestonly clientdistro=el7 mdsdistro=el7 ossdistro=el7 \
 mdsfilesystemtype=ldiskfs mdtfilesystemtype=ldiskfs \
   ostfilesystemtype=ldiskfs testlist=sanity envdefinitions=ONLY=17m

Change-Id: I1a80a80f6646984fbdec03d65579e735f6fac7c5
Signed-off-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>