LU-6654 osd: Add osd_iobuf concurrency control
The osd_iobuf struct in osd_thread_info is used by a thread
other than the one which owns the osd_thread_info.

The existing concurrency control (checking dr_numreqs) has
a race condition where dr_numreqs is decremented, which
opens the path for the other thread to eventually free
osd_thread_info, and then the first thread calls wakeup
on a wait queue which is inside the now freed
osd_thread_info.

This allocates the iobuf separately from osd_thread_info
and uses a spinlock around the decrement & wakeup and the
freeing operation, which prevents the freeing operation
from happening until the wakeup has happened.

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: If9b2f66e69e227d00a84660c66b856d29c7ab323