LU-6656 qmt: Fix non-atomic allocation under lock
In qmt_glimpse_lock:
While the resource lock (a spin lock) is held, it does

OBD_ALLOC_PTR(work);

This allocation is not atomic, so it can potentially sleep.
Sleeping with a spinlock held is not a good idea in
general, and (per Andreas) it is specifically not allowed
for GFP_NOFS allocations.

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: I4ed2cd5e127bb2aa47235416ff1f97b9ac7826bc