LU-6485 libcfs: distinguish kernelcomm by uid/pid/rfd
Currently, different lustre_kernelcomm are distinguished by
"uid" and "group", which is insufficient. For instance,
the "group" for HSM case will always be KUC_GRP_HSM(0x02),
and the lustre_kernelcomm(s) will be regarded to be the same
if they are created by the same user.

This patch will also compare the current pid and read fd(rfd)
to distinguish it from others kernelcomm in the same process
(by rfd) or from different processes (by pid).

Change-Id: Iaf5e9fbb9bad2540f51da2c4fd9c4047640d0877
Signed-off-by: Hongchao Zhang <hongchao.zhang@intel.com>