LU-6679 ldlm: do not set cbpending for group locks
Currently we will set CBPENDING on the group lock if there's another
group lock request on the same file but with a different gid.

Having the CBPENDING flag set will make ldlm_lock_match to fail and
the result is two group locks granted on the same resource.

This is essentially a missed case from LU-6368

Signed-off-by: Li Dongyang <dongyang.li@anu.edu.au>
Change-Id: I7af89e957528b3ed9771d86243ac8271084ee81f