LU-6356 tgt: handle sec context requests properly
In case of requests of type SEC_CTX_INIT, SEC_CTX_INIT_CONT
and SEC_CTX_FINI, no export is associated with the request
yet, because these requests are sent before *_CONNECT requests.
So the normal *_common_slice handlers will not be called,
given there is no reference to the target.

So add a special case for handling of SEC_CTX_INIT, SEC_CTX_INIT_CONT
and SEC_CTX_FINI requests in tgt_handle_request0() and
tgt_request_handle(), in order to directly send the reply.

Also fix error in sanityn.sh and see if it helps with Maloo.

Test-Parameters: fortestonly testlist=sanityn
Signed-off-by: Sebastien Buisson <sebastien.buisson@bull.net>
Change-Id: I08cb42a8f0eaa45b5d3c5b8052d89ae749c14d61