LU-6179 llite: Implement lock ahead
Lock ahead is a new feature allowing userspace to request
extent locks in advance of the IO which will use them.
These locks are not expanded beyond the size requested by
userspace.  Their use is intended to address lock
contention between multiple clients resulting from lock
expansion.  They should allow optimizing various IO
patterns.

Asynchronous glimpse locks are a speculative version of
glimpse locks, and already implement the required behavior.
Therefore, this patch takes the CEF_AGL flag and renames it
to CEF_SPECULATIVE.  (The same change is made for derived
flags at lower layers.)

There is no change to the behavior the flag invokes, but it
is now shared between AGL and lock ahead requests.

Additionally, lock ahead creates extent locks in advance of
IO, and so breaks the assumption that the holder of the
highest lock knows the current file size.

So we also modify the ofd_intent_policy code to glimpse
every PW lock and merge their LVBs.

Currently, lock ahead lock requests are always
non-blocking.  The hope is to change this in a future
version, but for now it avoids a number of LDLM deadlocks.

LDLM_FL_BLOCK_NOWAIT was only implemented for group locks,
so this patch also makes the changes necessary to allow it
for standard extent locks.

TODO:
Add tests - Tests file is present but is just a template.
(Suggestions for test cases are welcome.)
Other minor additions

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: I1e80286f54946a0df08b19b1339829fcfd1117e7