LU-6766 build: add support for arm64
Expand lustre to support the ARM64 platform.

The architecture is reported as aarch64 so add that support to
lustre-build-linux.m4.

For the Red Hat/Fedora packaging, strip the aarch64 suffix for the
required kernel.

Add definitions for lstat_t, lstat_f, fstat_f and fstatat_f for aarch64
architecture in lustre/include/lustre/lustre_user.h.

Change-Id: I2d4a9e7be4d4c32ab99da802f1b2318764e4386e
Signed-off-by: Arnaud Guignard <arnaud.guignard@cea.fr>