LU-3534 osp: transfer updates with bulk RPC
Send update request with bulk RPC(iovec), and update request
will be split into multiple buffers, so OSP does not need to
allocate big contingous buffer to store the update request.

If the request is allowed to be sent during recovery,
and it is not timeout yet, then we do not need to resend it
in the final stage of recovery. Otherwise the resend will
cause the bulk request to resend the request with different
mbit, but same xid, and on the remote server side, it will
drop such request with same xid, so it will never get the
bulk in this case.

Signed-off-by: wang di <di.wang@intel.com>
Change-Id: I0dd0c5fa66a7e71b8d4515571648abdf1a3d23a1