LU-6801 lfs: add --data option to HSM migrate
Add --data to "lfs migrate -H" to let the user pass some options to
the copytool. This could be used to pass options such as priority,
transfer throttle, ... Added a test for it.

The format of the data is JSON, because internally lfs is passing the
other parameters as JSON, so both parameters are merged.

Signed-off-by: frank zago <fzago@cray.com>
Change-Id: I97d3a34127365f08f29bfc8a96921f46da5d7e9e