CORRAIDZ-5 utils: remove libzfs_load_module() call
Remove libzfs_load_module() call from the OSD.  The function is no
longer exported by libzfs2 in ZFS master, as it was intended as an
internal function to begin with.  libzfs_init(), which is already
being used to set up the OSD, loads the module if necessary.

The intent of the call to libzfs_load_module apparently was to make
zfs_init() fail silently at sites using ldiskfs, but fail with an
error message if ZFS initialization failed.  It's not clear that
the code accomplished that anyway.

Handling such error conditions will be dealt with in a separate ticket
as it's a larger concern.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>
Change-Id: I6828e0aa8edbd318f5f5f820828d635ee060acb8