LU-6862 build: Make dist of contrib's contents consistent
Authors and reviewers of patches adding things to the contrib
subdirectory have been inconsistent with the inclusion of the
contrib tree's contents in the distribution tarball.

This patch fixes that by removing all of the Makefiles from the
contrib tree, and instead just listing contrib in EXTRA_DIST
at the top level.  That will recursively include the entire
contrib directory in the distribution tarball.  This way
there is little oportunity for inclusion of files under
contrib being inconsistent.

Change-Id: If6b92ed069a3fbeb1d0a8c8c0488180ee69597f3
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>