LU-6851 lnet: revert LU-6060 patch
Revert "LU-6060 lnet: set downis to 1 if there's no NI for remote net"

The logic here (as I understand it) is as follows:
1. ping information is returned to the pinger
2. for each route on the gateway on the pinger
2a. walk through all the NIs which the router returns
to a max of 16
3. for each NI if any of its nids are down then consider
that router down unless (4) evaluates to true
4. if the NI status is up then make sure that this NI is
on the same network as the destination net defined in the
route. If so then router is up, since we're able to get to
the destination
5. if you finish going through all the NIs and none of them
are down and none of them are on the same network as the
destination net defined in the route,
5a. then consider the router down if the number of hops for
this router is exactly 1 (which it is by default)

Problem is that there are two scenarios where the check
in (6) will evaluate to true:
1. This is a multihop system where the destination net
is not on the first hop and the # of hops is not defined
in the route (or it's wrong IE: 1)
2. LNet on the router came up without the interface which
is on the destination network (case in 6060).

From the point of view of the pinger, there is no way to
distinguish between these two scenarios without knowing whether
you're dealing with multihop route or not.

The assumption (as far as I can tell) which was made in the fix,
is that if you're in a multihop system then you'll always define
the number of hops in the route to > 1. If you don't define the number
of hops (then you default to 1) and therefore the first hop has to
have an interface on the destination network.

However since the # of hops is not a mandatory parameter the 6060
patch will cause problems to already existing systems since it
changes the behavior by which a route is determined down.

This reverts commit 749dc54622b2c3267c6c97eb770702b437a7897d.

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: I1a28a35a4edc2437cf95cb9d455e59c8102736fa