LU-6851 lnet: Default number of hops to -1
The logic in lnet_parse_rc_info() is as follows:
1. ping information is returned to the pinger
2. for each route on the gateway on the pinger
2a. walk through all the NIs which the router returns
to a max of 16
3. for each NI if any of its nids are down then consider
that router down unless (4) evaluates to true
4. if the NI status is up then make sure that this NI is
on the same network as the destination net defined in the
route. If so then router is up, since we're able to get to
the destination
5. if you finish going through all the NIs and none of them
are down and none of them are on the same network as the
destination net defined in the route,
5a. then consider the router down if the number of hops for
this router is exactly 1 (which it is by default)

Problem is that there are two scenarios where the check
in (5) will evaluate to true:
1. This is a multihop system where the destination net
is not on the first hop and the # of hops is not defined
in the route (or it's wrong IE: 1)
2. LNet on the router came up without the interface which
is on the destination network (case in LU-6060).

From the point of view of the pinger, there is no way to
distinguish between these two scenarios without knowing whether
you're dealing with multihop route or not.

The assumption (as far as I can tell) which was made in the fix for
LU-6060 is that if you're in a multihop system then you'll always
define the number of hops in the route to > 1. If you don't define
the number of hops (then you default to 1) and therefore the first
hop has to have an interface on the destination network.

However since the # of hops is not a mandatory parameter the 6060
patch will cause problems to already existing systems since it
changes the behavior by which a route is determined down.

To fix this case the # of hops now defaults to -1 if no hop count
is specified. When executing the logic in lnet_parse_rc_info()
there is no longer an assumption that the default hop count is 1.
If the hop count is 1 then it must've been explicitly set by the
user.

Furthermore, to maintain the established routing behabvior, when
selecting a route in LNetDist() or lnet_compare_routes() routes with
default hop count of -1 are delt with as if the hop count is 1.

This way the route selection algorithm will behave in an identical
manner as it does today.

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: I1a28a35a4edc2437cf95cb9d455e59c8102736fa