LU-6851 lnet: Ignore hops of not explicitly set
Since the # of hops is not a mandatory parameter the LU-6060
patch will cause problems to already existing systems since it
changes the behavior by which a route is determined down.

To fix this case the # of hops now defaults to UINT_MAX if
no hop count is specified. When executing the logic in
lnet_parse_rc_info() there is no longer an assumption that the
default hop count is 1. If the hop count is 1 then it must've
been explicitly set by the user.

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: I1a28a35a4edc2437cf95cb9d455e59c8102736fa