LU-6928 ptlrpc: Do not pack min XID for connect req
Do not pack min XID for connect request, because all of
unreplied request may not in the sending/delay list at the
moment, for example replay request which is just timeout.

And if it packs the wrong min(bigger than it should be) XID
to the Server, and server will delete some reply data, which
client even do not know if the server receives, so the client
might resend the request, but server already delete the reply
data, so the request will be executed again.

Signed-off-by: wang di <di.wang@intel.com>
Change-Id: I43698ec64199319091f4840636be7b38a7ac6221