LU-5319 tests: testcases for multiple modify RPCs feature
This patch creates new testcases in the Auster test suite to
test the support of multiple modify RPCs in flight feature.

Two new OBD_FAIL codes are added to allow several failure occurences
of the reint requests or replies. This is needed because the current
fail checks impose the OBD_FAIL_ONCE flag.

Added testcases are :
- sanity         245
- conf-sanity    90a 90b 90c 90d
- replay-single  102a 102b 102c 102d

This patch also removes replay-single tests 53a and 53b
from the ALWAYS_EXCEPT list.

Signed-off-by: Gregoire Pichon <gregoire.pichon@bull.net>
Change-Id: I3cd233d5ef67a696e830c77b154dcc01520bb3d2