LU-6932 utils: bound check the index in llog_reader
A corrupted file can crash the llog_reader utility, if the index read
is out of bound. Add an extra check to catch that case.

Also fixed a typo, and replaced a printf with llapi_error, since
errors should go to stderr, and updated its error message with better
information.

Signed-off-by: frank zago <fzago@cray.com>
Change-Id: I16abe55081cccc8363f6b31c57710ea154e9037e