LU-5030 utils: create a libcfs function to handle tunable paths
With lustre going to sysfs in the future but at the same time
many people using older versions of Lustre adds a requirement
of suporting tunables for both the procfs and sysfs backends.
Here we create a new library function that autoexpands the
path and test for the presence of one or the other and sets
or gets the tunable.

Change-Id: I7fbb7348e009cb9f77ba9847aeb4879dc302e4ee
Signed-off-by: Dmitry Eremin <dmitry.eremin@intel.com>
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Signed-off-by: Wang Chao <chao.ornl@gmail.com>