LU-5092 nodemap: handle config changes while mid-flight
This adds the ability to detect when the nodemap configuration has
changed between messages for very large configurations. If the
configuration is over 1MB (about 25k entries), it needs to be sent
in multiple messages. If the configuration changes in-between these
messages, then the client needs to abort the previous get config
operation and restart.

Signed-off-by: Kit Westneat <kit.westneat@gmail.com>
Change-Id: I2bf6e4bade947fea8331c2922d9c33ab49100577