LU-6214 llite: tar --xattr restore fails for HSM released files.
If you create a file, archive and release it, it keeps only a
link and all information in xattr. If you tar the file
with --xattr you will store the same striping information and link
information in the tar. If you delete the file, the file and archive
state does not make sense. Now if you restore the file using tar
with xattr having the RELEASED flag turned on, then it is not correct
because this is a new file. So I am ignoring the HSM xattr and masking
out the "RELEASED" flag for the files, which are not archived.
Added testcase for the same.

Change-Id: I0ca8636bf508211d63ba796c3b756c9cb4b42d48
Signed-off-by: Aditya Pandit <panditadityashreesh@yahoo.com>