LU-6356 tgt: add handlers for SEC_CTX_* requests
When dealing with security context refresh requests, an export is
already associated with the request. So the normal *_common_slice
handlers are called.
We then have to have handlers for SEC_CTX_INIT, SEC_CTX_INIT_CONT
and SEC_CTX_FINI requests in *_common_slice for all target types.

This patch reintroduces the handlers that were removed by patch
http://review.whamcloud.com/14041, plus adds a new one for ofd.

Signed-off-by: Sebastien Buisson <sebastien.buisson@bull.net>
Change-Id: Ifadaff9d7f07cf4c6a75cd1f5054378aaca27c26